+++ clients/fides-js/src/lib/consent-constants.ts
@@ -25,6 +25,12 @@ export const FIDES_OVERRIDE_OPTIONS_VALIDATOR_MAP: {
     validationRegex: /^(true|false)$/,
   },
   {
+    fidesOption: "fidesDisableBanner",
+    fidesOptionType: "boolean",
+    fidesOverrideKey: "fides_disable_banner",
+    validationRegex: /^(true|false)$/,
+  },
+  {
     fidesOption: "fidesString",
     fidesOptionType: "string",
     fidesOverrideKey: "fides_string",